Introduction to tables
I had to work on commenting everything out that was already in the file
just as Colt did in his videos, and edit comments within the commented
out block so that it wouldn’t cause any issues. After that I added in a
table with a border, 2 rows, and 2 columns each.