Correction of Files, and expanding data in tables as per Lesson
I forgot that I have a project folder in Atom only selected on
Section3, so I edited section 3 like I was trying to do for section 4.
Beyond that the actual edits to section 4 itself was part of the lesson
plan, which he expands the data in the table to reflect his dogs, and
shows the use of <thead> and <tbody> tags as well as <th>. This is good
to know how to do so you can semantically build out your HTML file
properly.