Added labels to our form
Here we added labels to our form. I know that we used them, have used
them in the past however a good thing to note is that we put our inputs
in our label to connect them or use for=“[id]” where id is the id of
the input you are trying to connect. That is something I didn’t know.