Image Gallery Code Along
This was a series of videos that had us coding along to create this
nice photo gallery. I tried to stay one step ahead of him and try to
predict what he was going to do, and ended up finishing before he did.
I like this photo gallery and it might be useful for something else in
the future.