Conditional Exercise
I was given the task of differentiating between different ages. We had
to make sure someone could enter a bar and whether or not they could
drink after they entered. We were given 2 bonuses, the first being
making sure the input is not negative as you cannot be negative years
old, and the second being to check to see if someones age is a perfect
square and point that out to the user if it is.